Separated mapping and read/write logic from watcher
